Extract shared API key and request helpers to test-utils
- Move getTestDataKey, createTestApiKeyToken, apiRequest to #test-utils
  for reuse across admin API and API key test suites
- Replace large inline EventWithCount literals with testEventWithCount()
- Remove duplicate getTestDataKey from api-keys.test.ts
- Remove duplicate apiRequest/createTestApiKey from admin-api-events.test.ts

Reduces jscpd clones from 1230 to 1224 (11.16% → 11.09%).
